+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Persistence.Core/Executors/IDbExecutor.cs
@@ -1,5 +1,7 @@
-﻿using Altinn.AccessMgmt.Persistence.Core.Helpers;
+﻿using Altinn.AccessMgmt.Persistence.Core.Definitions;
+using Altinn.AccessMgmt.Persistence.Core.Helpers;
 using Altinn.AccessMgmt.Persistence.Core.Models;
+using Altinn.AccessMgmt.Persistence.Core.QueryBuilders;
 
 namespace Altinn.AccessMgmt.Persistence.Core.Executors;
 
@@ -60,5 +62,20 @@ public interface IDbExecutor
     /// <param name="cancellationToken">CancellationToken</param>
     /// <returns></returns>
     Task<IEnumerable<T>> ExecuteMigrationQuery<T>(string query, CancellationToken cancellationToken = default)
+    where T : new();
+
+    /// <summary>
+    /// Performs a bulk ingest operation by importing a list of entities into the database.
+    /// NOT TO BE USED LIGHTLY!
+    /// </summary>
+    /// <param name="data">The list of entities to import.</param>
+    /// <param name="definition">DbDefinition</param>
+    /// <param name="queryBuilder">IDbQueryBuilder</param>
+    /// <param name="batchSize">Batch size (default: 1000)</param>
+    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
+    /// <returns>
+    /// A task that represents the asynchronous operation. The task result contains the number of entities successfully ingested.
+    /// </returns>
+    Task<int> Ingest<T>(List<T> data, DbDefinition definition, IDbQueryBuilder queryBuilder, int batchSize = 1000, CancellationToken cancellationToken = default)
     where T : new();
 }
