+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Persistence.Core/Executors/PostgresDbExecutor.cs
@@ -1,17 +1,21 @@
 ﻿using System.Data;
+using Altinn.AccessMgmt.Persistence.Core.Definitions;
+using System.Reflection;
 using Altinn.AccessMgmt.Persistence.Core.Models;
 using Altinn.AccessMgmt.Persistence.Core.Utilities;
 using Altinn.Authorization.Host.Database;
 using Altinn.Authorization.Host.Startup;
 using Microsoft.Extensions.Logging;
 using Npgsql;
+using NpgsqlTypes;
+using Altinn.AccessMgmt.Persistence.Core.QueryBuilders;
 
 namespace Altinn.AccessMgmt.Persistence.Core.Executors;
 
 /// <summary>
 /// Responsible for executing SQL commands and queries.
 /// </summary>
-public class PostgresDbExecutor(IAltinnDatabase databaseFactory, NpgsqlDataSource connection, IDbConverter dbConverter) : IDbExecutor
+public class PostgresDbExecutor(IAltinnDatabase databaseFactory, IDbConverter dbConverter) : IDbExecutor
 {
     /// <summary>
     /// Logger instance for logging database configuration messages.
@@ -19,8 +23,17 @@ public class PostgresDbExecutor(IAltinnDatabase databaseFactory, NpgsqlDataSourc
     private static ILogger Logger { get; } = StartupLoggerFactory.Create(nameof(PostgresDbExecutor));
 
     private readonly IAltinnDatabase _databaseFactory = databaseFactory;
-    private readonly NpgsqlDataSource _connection = connection;
     private readonly IDbConverter _dbConverter = dbConverter;
+
+    /// <summary>
+    /// Temp Connection Exposure for Ingest
+    /// </summary>
+    /// <param name="sourceType">SourceType (App, Migrate)</param>
+    /// <returns></returns>
+    public NpgsqlConnection GetConnection(SourceType sourceType)
+    {
+        return _databaseFactory.CreatePgsqlConnection(sourceType);
+    }
 
     /// <summary>
     /// Executes a query and maps the result to objects of type T.
@@ -140,5 +153,112 @@ public class PostgresDbExecutor(IAltinnDatabase databaseFactory, NpgsqlDataSourc
         cmd.CommandText = query;
         conn.Open();
         return _dbConverter.ConvertToObjects<T>(await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, cancellationToken));
+    }
+
+    /// <inheritdoc />
+    public async Task<int> Ingest<T>(List<T> data, DbDefinition definition, IDbQueryBuilder queryBuilder, int batchSize = 1000, CancellationToken cancellationToken = default) 
+    where T : new()
+    {
+        var type = typeof(T);
+
+        using var conn = _databaseFactory.CreatePgsqlConnection(SourceType.Migration);
+        if (conn.State != ConnectionState.Open)
+        {
+            conn.Open();
+        }
+
+        var dt = new DataTable();
+
+        // Use a simple query to get a sample structure.
+        var dataAdapter = new NpgsqlDataAdapter($"SELECT * FROM {queryBuilder.GetTableName(includeAlias: false)} LIMIT 10", conn);
+        dataAdapter.Fill(dt);
+        dt.Clear();
+
+        var columns = new Dictionary<string, (NpgsqlDbType Type, PropertyInfo Property)>();
+
+        foreach (DataColumn c in dt.Columns)
+        {
+            if (!definition.Properties.Exists(t => t.Name.Equals(c.ColumnName, StringComparison.CurrentCultureIgnoreCase)))
+            {
+                continue;
+            }
+
+            columns.Add(c.ColumnName, (GetPostgresType(c.DataType), definition.Properties.First(t => t.Name.Equals(c.ColumnName, StringComparison.CurrentCultureIgnoreCase)).Property));
+        }
+
+        string columnStatement = string.Join(',', columns.Keys);
+        using var writer = await conn.BeginBinaryImportAsync($"COPY {queryBuilder.GetTableName(includeAlias: false)} ({columnStatement}) FROM STDIN (FORMAT BINARY)", cancellationToken: cancellationToken);
+        writer.Timeout = TimeSpan.FromMinutes(10);
+        int batchCompleted = 0;
+        int completed = 0;
+        foreach (var d in data)
+        {
+            writer.StartRow();
+            foreach (var c in columns)
+            {
+                try
+                {
+                    writer.Write(c.Value.Property.GetValue(d), c.Value.Type);
+                }
+                catch (Exception ex)
+                {
+                    // Replace with proper logging.
+                    Console.WriteLine($"Failed to write data in column '{c.Key}' for '{definition.ModelType.Name}'. Trying to write null. " + ex.Message);
+                    try
+                    {
+                        writer.WriteNull();
+                    }
+                    catch
+                    {
+                        Console.WriteLine($"Failed to write null in column '{c.Key}' for '{definition.ModelType.Name}'.");
+                        throw;
+                    }
+                }
+            }
+
+            completed++;
+            if (completed == batchSize)
+            {
+                batchCompleted++;
+                completed = 0;
+                Console.WriteLine($"Ingested {(batchCompleted * batchSize) + completed}");
+            }
+        }
+
+        Console.WriteLine($"Ingested {(batchCompleted * batchSize) + completed}");
+        writer.Complete();
+
+        return (batchCompleted * batchSize) + completed;
+    }
+
+    private NpgsqlDbType GetPostgresType(Type type)
+    {
+        if (type == typeof(string))
+        {
+            return NpgsqlDbType.Text;
+        }
+
+        if (type == typeof(int))
+        {
+            return NpgsqlDbType.Integer;
+        }
+
+        if (type == typeof(DateTimeOffset))
+        {
+            return NpgsqlDbType.TimestampTz;
+        }
+
+        if (type == typeof(Guid))
+        {
+            return NpgsqlDbType.Uuid;
+        }
+
+        if (type == typeof(bool))
+        {
+            return NpgsqlDbType.Boolean;
+        }
+
+        Console.WriteLine($"Type converter not found for '{type.Name}'");
+        return NpgsqlDbType.Text;
     }
 }
