--- src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Persistence.Core/Helpers/IDbIngest.cs
@@ -1,17 +1,0 @@
-﻿namespace Altinn.AccessMgmt.Persistence.Core.Helpers;
-
-/// <summary>
-/// Interface for ingesting data into the database
-/// </summary>
-/// <typeparam name="T"></typeparam>
-public interface IDbIngest<T> 
-    where T : class, new()
-{
-    /// <summary>
-    /// Ingest data into the database
-    /// </summary>
-    /// <param name="data">Data to ingest</param>
-    /// <param name="cancellationToken">CancellationToken</param>
-    /// <returns></returns>
-    Task<int> Ingest(List<T> data, CancellationToken cancellationToken = default);
-}
